Implement ImageSelector, Remove FileSelector
